+++ changelog.md
@@ -3,6 +3,7 @@
 ## Upcoming
 * Added Red block outline with wrong tool
   * If the tool in hand will make the block drop no items then the block outline will be green tinted
+* Added water push strength and push even if there are no blocks nearby
 * Berry bushes patches now have fewer bushes and no longer generate all fully grown
 * Slightly increased coal, gold, iron and lapis generation
 * Huge mushrooms now generate less in Dark Oak Forests